+++ website/src/components/Chat/ChatListDesktop.tsx
@@ -4,7 +4,7 @@ import { memo } from "react";
 import { ChatListBase } from "./ChatListBase";
 
 // eslint-disable-next-line react/display-name
-export const ChatList = memo(() => {
+export const ChatListDesktop = memo(() => {
   return (
     <Box pe="260px" display={{ base: "none", md: "block" }}>
       <ChatListBase></ChatListBase>
